ART
Had the basic layout of the application written down and I've got these so far.

Berry-nebula is the color pallete I found and enjoyed. Would be interested in easily changing color pallets in game.

Blends are the 3D models for the 3d world portview.

Found a good looking pikmin font.

Made these sprites with asesprite. Hoping to set everything up to make it as easy as possible for keegan to update the visuals so they don't look like doodoo