Panning
Got the camera to pan as I had imagined. The camera starts on the right side of the art, with a distance from its center between 4-8 units. Its panning goal is negative of this distance. A y level for both the start and end is random between 8 - 14 units. This gives a random diagnol affect.

Z is caluclated on a bunch of math I had to work out on paper