Camera Fade
Camera now fades out when it reaches the end of its path. Reaching the end starts the fadeout animation, starts a timer that when finished will signal the SetCamera function, which fades in at the end.

Side Note: the reason Im checking if the remaining distance from currentPosition to endposition <= .2f. Since I want the camera to go left OR right, I need to check the absolute value of the difference since either could be positive/negative. So I can't just check if its <= 0. And == anything wont work cause of the impercision and the fact its a process function. so <=.2f seemed like a good buffer.