Camera Refactor
After adding a 2nd room, it was obvious that the camera was gonna have to work in a couple of different fasions. So I've moved the camera logic to its own class outside of portview so that I can have different cameras

Right now my only camera is the one that pans across the art, but I'll also need some that sit still and rotate. Also made the cameras work with art being anywhere, so the range it pans isnt just hardcoded values