Uploading Art Configurations
Really happy with this. Basically spent a little time over the weekend getting set up with an AWS account and refamilizring myself with the API and lambda stuff.

Then today I started fresh with a new api, lambda, and S3 bucket and worked my way to being able to make a request from godot, to the api to upload all the currently loaded art configurations. Thats most of the additions to ArtIcon.cs. One thing I'd like to do is see if I can overload the searalizer function for artIcon. *shrug*

Had to learn how to sent http requests in godot. Pretty simple! Biggest hurtle was learning to use the requester class instead of the client. One good example from the documentation and I was pretty set.

Was having trouble getting the request to path through both the api method integration and lambda integration in the correct way. Finally found a good source for the specific template language that aws uses for the the lambda mapping. Just had to tweak the example to include what I needed and bam.

The rest of anything I had to do and learn is related to the aws code so I won't go into that here