Camera Fix
when changing the visible scene in the viewport, the camera wasn't correctly changing. So if I switched from the spotlight room to the kitchen, it would keep the panning camera and look wrong. But I figured out that I need to change which camera is "current" by using the MakeCurrent() function. That fixed up my issue