StarRating edit
Woo! Ok got the star rating working as I'd imagined. I decided to try using the Control nodes Focus Neighbors instead of implemnting a custom UI traverser like I ended up doing for the art icons. That wasnt difficult to figure out, just some trial and error with grabbing and setting the focus.

Star rating is simple, it works in .5f increments. First we fill however many stars. So a 2.5f score will have two stars filled. Then 5-2 stars will be set as empty. Then we check if the score has a half and if so set that star to a halfStar.

Needed to change data types in the deserializer but that was it in terms of whats already been implemented for the artIcons