Local Saving and Detail Edits
I can now change the title of any artpiece and save the data locally instead of remotely. Took a bit to get the controls right, had to edit some of the builtin UI controls. For example, esc was set to have nodes lose focus but I just needed it to close the keyboard input.