Star Highlighting
I tried creating a seperate highlight file for the stars like I did with the other details, but I didnt like the look of highlighting each star. So I opted for a simple rectangle highlight. But since this is made of multiple sprite (unlike title size or id) I figured I'd just make a seperate sprite in the scene that becomes visible or not visible. UNLIKE the others where I basically swap the texture of the background to a highlighted version