Valid ArtId Not Overwritten
During testing I noticed that when I deleted all details file, godot took forever to import? This was because my load logic searched for a details file, and if not found would rename the art and the new details file.

Now instead we check if the art has a valid name already, and if so we just create the details file and we don't bother renaming the art since its already valid