HOLY SHIT THE OPTIMIZING
Spent alllllll day yesterday improving the applications ability to be performant. So after I got all the photos read in from another directory, things immedietly went badly XD. I pretty quickly found out that the VRAM was simply exploding from the photos being dynamically loaded in. Like my graphics card has 8gb of dedicated memory and with just 25ish images it was bumping up to 6gb, EVEN with the much lower quality icon images.

So I started getting into that by changing up the IconCollections init, making sure that details and lowrez images were available for all images at the start, but that init function was EXPLODING with functionality that frankly had nothing to do with the icon collection anymore and it was a mess to get right. So I took a step back and created the File Manager. Now at startup, the RootWindow calls for the Filemanager to do an Init, which will check that the 3 mandatory directories exist, and double check the following things
1. all art in ART should have a proper id for its name.
2. every art piece should have a details file, and if not one is created
3. every art piece should have a low rez image and one is created if not.

This means its pretty easy for duplicate/orphaned files to exist but thats a generally easy problem to tackle in the future. SO now all the logic that defines the specific directories Artlopedia needs is outside the iconCollection. This was also a good place for me to add some easy functions to get the correct directories in other parts of the application.

Once I moved all the functionality I felt was out of palce into FileManager, then I spent a BUNCH of time figuring out what to do about the crazy memory. I came across an immediate and big step examining the ImageTexture options and found the Compress function. This made the total amount of memory muuuuuch more reasonable, even with the high rez images. But even then 50+ images was absolutely fucking the application.

So I had to spend a bunch of time testing and monitoring the vram while trying to remove the textures from memory. Simply removing ArtIcons wasnt working, nullifying textures or trying to dispose them failed. Eventually I found a good combination of first finding all references to the textures (kept them contained to just artIcon thank you very much) to null them, THEN calling the OS garbage collector and that did the trick